requirements.txt
bidict==0.21.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
eventlet==0.31.1
Flask==2.0.1
Flask-SocketIO==5.1.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.7.1
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
requests==2.26.0
six==1.16.0
twilio==6.63.0
urllib3==1.26.6
Werkzeug==2.0.1
python-dotenv~=0.19.0